refactor: restructure package for profiler classes
