+++ src/main/java/dev/tori/runtimeprofiler/profiler/ProfileEntry.java
@@ -19,7 +19,7 @@
  * THE SOFTWARE.
  */
 
-package dev.tori.runtimeprofiler;
+package dev.tori.runtimeprofiler.profiler;
 
 import dev.tori.runtimeprofiler.util.Stopwatch;
 import dev.tori.runtimeprofiler.util.UnitUtil;
